requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
djangorestframework-jsonapi==2.6.0
inflection==0.3.1
psycopg2==2.7.5
pytz==2018.6
six==1.11.0
